Create LICENSE
Hey! I'm just working on my final degree project "Autonomy of Follow me cars" and I would like to use part of your code to create a tensorflow lite model to find airplanes near the car. If it is not problem to you I would like to add a open source license to your project to use part of it. Of course I will mentioned you as a reference in the code example. Thank you!